+++ ginga/mplw/ImageViewMpl.py
@@ -10,6 +10,7 @@ from io import BytesIO
 # Matplotlib imports
 import matplotlib
 #from matplotlib.path import Path
+import numpy as np
 
 from ginga import ImageView
 from ginga import Mixins, Bindings
@@ -166,7 +167,7 @@ class ImageViewMpl(ImageView.ImageViewBase):
 
         # Grab the RGB array for the current image and place it in the
         # matplotlib figure axis
-        data = self.getwin_array(order=self.rgb_order)
+        data = self.getwin_array(order=self.rgb_order, dtype=np.uint8)
 
         dst_x = dst_y = 0
 
@@ -201,7 +202,7 @@ class ImageViewMpl(ImageView.ImageViewBase):
 
         # Grab the RGB array for the current image and place it in the
         # matplotlib figure axis
-        arr = self.getwin_array(order=self.rgb_order)
+        arr = self.getwin_array(order=self.rgb_order, dtype=np.uint8)
 
         # Get the data extents
         x0, y0 = 0, 0
